backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
pymongo==4.6.0
dnspython==2.4.2
